requirements.txt
websocket-client
pymysql
pika
whisper
cryptography
bilibili-api-python==16.2.0